fixing line with JOBID
The cells are isolated, so we cannot use flux jobs and
see anything from a previous submit. We also cannot use
an environment variable to pipe in a JOBID. This fix will
instead run flux job last directly in the command to
properly show the job info resource spec and output. I also
added showing the user how to write to a custom temporary file.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>